src/requirements.txt
pycti==6.4.11
pyyaml~=6.0.2
requests~=2.32.3
validators==0.35.0
beautifulsoup4~=4.12.3
